requirements.txt
cloudinary
Flask
opencv-python
flask-cors
flask-restx
python-dotenv
referencing
requests
matplotlib
numpy==1.22.0
